\name{simulatedRankingData}
\docType{data}
\alias{simulatedRankingData}
\title{
Simulated ranking data
}
\description{
The \code{simulatedRankingData} dataset is a list that includes the following components: 

The ranking matrix, \code{y}, contains the ranking matrix. It consists of 500 rows and 4 columns, indicating the ranking positions. Each element in the matrix represents the rank assigned to an individual for a particular item.

The predictor matrix \code{x} in the dataset consists of 20 continuous explanatory variables. These variables are used for predicting the rankings.
}
\usage{data(simulatedRankingData)}
\references{
Albano, A., Sciandra, M., and Plaia, A. (2023): "A weighted distance-based approach with boosted decision trees for label ranking." \emph{Expert Systems with Applications}.
}
\keyword{datasets}
